fix(api)!: report the real rule count, and drop five unused public types
`PolicyError::TooManyRules` narrowed `got`/`max` to `u8`, so a policy of
300 rules
reported "egress policy has 255 rules, over the 16-rule cap" — a false
statement about
the caller's own input on a security-control path, since
`EgressPolicy::allow` is an
uncapped push. Back to `usize`, and `max` reads `MAX_POLICY_RULES`
instead of a
hardcoded fallback that would go stale if the cap moved.

FilterAxis, FilterMode, RuleStatus, RuleProtocol and OutputFormat had no
callers.
FilterAxis existed only to define the constants it was meant to replace,
and its
insertion had orphaned `FILTER_TGID`'s doc comment onto the enum.

+++ crates/probes-loader/src/egress.rs
@@ -25,9 +25,9 @@ pub enum PolicyError {
     /// More allow-rules than the kernel `POLICY` map holds: the requested count and the cap.
     TooManyRules {
         /// The number of rules the caller supplied.
-        got: u8,
+        got: usize,
         /// The fixed cap ([`crate::MAX_POLICY_RULES`]).
-        max: u8,
+        max: usize,
     },
 }
